adds formatters
