+++ src/ui/mod.rs
@@ -1,5 +1,6 @@
 pub mod action_menu;
 pub mod container_list;
+pub mod formatters;
 pub mod help;
 pub mod input;
 pub mod log_view;
